Fix #4526 / #4528: revert IRevisioned to int (V8) + add ILongVersioned (long) support
Marten 9.0-alpha forked IRevisioned to long Version. jasperfx#348 keeps the
canonical JasperFx.IRevisioned at int and adds JasperFx.ILongVersioned (long) for
documents projected from a MultiStreamProjection whose Version is the global
event sequence number. This consumes both and keeps the bigint mt_version column.

Production:
- Delete Marten.Metadata.IRevisioned (long); alias IRevisioned -> JasperFx.IRevisioned
  (int) and ILongVersioned -> JasperFx.ILongVersioned (long) in the shared dedupe file.
- VersionedPolicy recognizes both IRevisioned and ILongVersioned -> UseNumericRevisions,
  mapping Metadata.Revision.Member.
- MetadataColumn<T>.Member gains a virtual IsAcceptableMemberType hook; RevisionColumn
  overrides it to accept int (IRevisioned) or long (ILongVersioned). Column stays bigint.
- DocumentRevisionBinder builds a member-type-aware setter: for an int member it
  downcasts the loaded long (documented overflow caveat for huge event sequences);
  for a long member it assigns directly.
- DocumentSessionBase.storeEntity gains an ILongVersioned case beside IRevisioned;
  revision flows as long internally (int upcasts).

Tests/examples: the revision-counter docs that had adopted the alpha long signature
revert to int Version (all are single-stream revision docs). Added
long_versioned_revisioning with an ILongVersioned doc whose Version > Int32.MaxValue
round-trips through the bigint column without truncation.

Docs: concurrency.md documents IRevisioned (int) vs ILongVersioned (long) and the
int-overflow caveat for MultiStreamProjection-derived documents.

Full solution builds clean (net9/net10). Revision/concurrency tests: DocumentDbTests
numeric + long revisioning 24/0; EventSourcing FetchForWriting/compacting/last-good 227/0.

Closes #4526 and #4528.

+++ src/EventSourcingTests/Aggregation/stream_compacting.cs
@@ -362,7 +362,7 @@ public class Letters : IRevisioned
     public int BCount { get; set; }
     public int CCount { get; set; }
     public int DCount { get; set; }
-    public long Version { get; set; }
+    public int Version { get; set; }
 }
 
 public class LetterCounts: IRevisioned
@@ -372,7 +372,7 @@ public class LetterCounts: IRevisioned
     public int BCount { get; set; }
     public int CCount { get; set; }
     public int DCount { get; set; }
-    public long Version { get; set; }
+    public int Version { get; set; }
 }
 
 public class LetterCountsByString: IRevisioned
@@ -382,7 +382,7 @@ public class LetterCountsByString: IRevisioned
     public int BCount { get; set; }
     public int CCount { get; set; }
     public int DCount { get; set; }
-    public long Version { get; set; }
+    public int Version { get; set; }
 }
 
 public partial class LetterCountsByStringProjection: SingleStreamProjection<LetterCountsByString, string>
